+++ envs/dev/main.tf
@@ -26,14 +26,14 @@ EOF
   base_user_data_suffix = <<-EOF
       - name: root
         lock_passwd: false
-        hashed_passwd: ${var.root_password_hash}
+        hashed_passwd: ${module.config.root_password}
         ssh_authorized_keys:
-          ${indent(6, join("\n", formatlist("- \"%s\"", module.common.root_ssh_keys)))}
+          ${indent(6, join("\n", formatlist("- \"%s\"", module.config.ssh_keys)))}
       - name: jderose
         groups: sudo
         shell: /bin/bash
         ssh_authorized_keys:
-          ${indent(6, join("\n", formatlist("- \"%s\"", module.common.jderose_ssh_keys)))}
+          ${indent(6, join("\n", formatlist("- \"%s\"", module.config.ssh_keys)))}
         sudo: ALL=(ALL) NOPASSWD:ALL
 
     package_update: false
@@ -57,14 +57,14 @@ EOF
     users:
       - name: root
         lock_passwd: false
-        hashed_passwd: ${var.root_password_hash}
+        hashed_passwd: ${module.config.root_password}
         ssh_authorized_keys:
-          ${indent(6, join("\n", formatlist("- \"%s\"", module.common.root_ssh_keys)))}
+          ${indent(6, join("\n", formatlist("- \"%s\"", module.config.ssh_keys)))}
       - name: jderose
         groups: sudo
         shell: /bin/bash
         ssh_authorized_keys:
-          ${indent(6, join("\n", formatlist("- \"%s\"", module.common.jderose_ssh_keys)))}
+          ${indent(6, join("\n", formatlist("- \"%s\"", module.config.ssh_keys)))}
         sudo: ALL=(ALL) NOPASSWD:ALL
 
     package_update: false
@@ -115,7 +115,7 @@ module "sdn" {
   source         = "../../proxmox-sdn"
   zone_id        = "homestak"
   vnet_id        = "vnet10"
-  peers          = ["10.0.12.124"]
+  peers          = [module.config.node_ip]
   subnet_cidr    = "10.10.10.0/24"
   subnet_gateway = "10.10.10.1"
 }
@@ -140,7 +140,7 @@ module "router" {
   source = "../../proxmox-vm"
 
   cloud_image_id    = module.cloud_image.file_id
-  proxmox_node_name = var.proxmox_node_name
+  proxmox_node_name = module.config.node
   vm_id             = 10000
   vm_name           = "router"
   vm_memory         = 1024